+++ src/main/java/otus/library/dao/GenreDao.java
@@ -4,8 +4,8 @@ import otus.library.domain.Genre;
 import java.util.List;
 
 public interface GenreDao {
-    Integer count();
+    Long count();
     void insert(Genre genre);
-    Genre getById(Integer id);
+    Genre getById(Long id);
     List<Genre> getAll();
 }
