+++ src/main/java/otus/library/dao/GenreDaoJdbc.java
@@ -15,7 +15,7 @@ public class GenreDaoJdbc implements GenreDao {
 
     private static class GenreMapper implements RowMapper<Genre> {
         public Genre mapRow(ResultSet resultSet, int i) throws SQLException {
-            Integer id = resultSet.getInt("id");
+            Long id = resultSet.getLong("id");
             String name = resultSet.getString("name");
             return new Genre(id, name);
         }
@@ -25,15 +25,15 @@ public class GenreDaoJdbc implements GenreDao {
         this.jdbc = jdbc;
     }
 
-    public Integer count(){
-        return jdbc.queryForObject("select count(*) from genre", Integer.class);
+    public Long count(){
+        return jdbc.queryForObject("select count(*) from genre", Long.class);
     }
 
     public void insert(Genre genre){
         jdbc.update("insert into genre (id, `name`) values (?, ?)", genre.getId(), genre.getName());
     }
 
-    public Genre getById(Integer id){
+    public Genre getById(Long id){
         return jdbc.queryForObject("select * from genre where id = ?", new Object[] {id}, new GenreDaoJdbc.GenreMapper());
     }
