+++ src/main/java/otus/library/domain/Author.java
@@ -1,17 +1,17 @@
 package otus.library.domain;
 
 public class Author {
-    private final Integer id;
+    private final Long id;
     private final String fname;
     private final String lname;
 
-    public Author(Integer id, String fname, String lname){
+    public Author(Long id, String fname, String lname){
         this.id = id;
         this.fname = fname;
         this.lname = lname;
     }
 
-    public Integer getId(){
+    public Long getId(){
         return this.id;
     }
